+++ reiche-backend/.gitignore
@@ -27,5 +27,3 @@ go.work.sum
 static/
 tmp/
 
-serve
-
