+++ reiche-backend/go.mod
@@ -2,12 +2,15 @@ module reiche
 
 go 1.23.2
 
-require github.com/julienschmidt/httprouter v1.3.0
+require (
+	github.com/cespare/xxhash v1.1.0
+	github.com/go-json-experiment/json v0.0.0-20240815175050-ebd3a8989ca1
+	github.com/julienschmidt/httprouter v1.3.0
+	zombiezen.com/go/sqlite v1.4.0
+)
 
 require (
-	github.com/cespare/xxhash v1.1.0 // indirect
 	github.com/dustin/go-humanize v1.0.1 // indirect
-	github.com/go-json-experiment/json v0.0.0-20240815175050-ebd3a8989ca1 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/ncruces/go-strftime v0.1.9 // indirect
@@ -17,5 +20,4 @@ require (
 	modernc.org/mathutil v1.6.0 // indirect
 	modernc.org/memory v1.8.0 // indirect
 	modernc.org/sqlite v1.33.1 // indirect
-	zombiezen.com/go/sqlite v1.4.0 // indirect
 )
