+++ reiche-backend/internal/handlers/handleutils/types.go
@@ -15,18 +15,3 @@ type ReqErr struct {
     Ctx ReqErrCode `json:"ctx"`
 }
 
-type ReicheFileType uint8
-const (
-    Image ReicheFileType = iota
-    Video
-    Object3D
-    Unreachable
-)
-
-type ReicheFile struct {
-    Path   string
-    Type   ReicheFileType
-    Ext    string
-    Hashed bool
-}
-
