--- reiche-backend/internal/handlers/router.go
@@ -1,10 +1,0 @@
-package handlers
-
-import "github.com/julienschmidt/httprouter"
-
-func NewReicheRouter() *httprouter.Router {
-    new_router := httprouter.New()
-    new_router.POST("/upload", UploadFile)
-    return new_router
-}
-
